Begin integrating other RPs into the suite
(In alphabetic order) Routinator, rpki-client and rpki-prover.

I say "begin" because only the VRP count and HTTP/rsync requests checks
are portable at the moment.

The other RPs more or less work. Routinator needs to be manually patched
because it has validations that prevent exactly what I'm trying to do
(RPT bug, not a Routinator one) and Prover takes a while to terminate
despite seeming to finish the validation early, for what appears to be
an over-sleeping thread. Worth some research.