Drop standard error of `$RP_BIN $RP_TEST`
While printing the version number of all relevant binaries in the
beginning is not a bad idea, some binaries don't seem to support -V,
so their `$RP_TEST` has to be --help. Printing --help in the beginning
is definitely too much noise to be a good idea.

So, for consistency, I was sending `$RP_BIN $RP_TEST`'s standard output
to /dev/null.

But it turns out rpki-client echoes -V into standard error,
so /dev/null that too.