Reduce for Routinator debugging
Despite applying some feedback, Routinator still refuses to print an
error message as a result of the failing HTTP requests.

This commit reduces the project as much as possible, so I can suggest
Martin to look at it and not waste too much of his time.

Also, he asked me to move the commandline args to routinator.conf.
Did it.

This is a temporal commit; I'll rollback it when the problem has been
found and corrected.